+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file. This change
 ## [Unreleased]
 ### Added
 - Native temp file and dir facilities ([#934](https://github.com/planck-repl/planck/issues/934))
+- Ability to open i/o streams on URI and JAR resources ([#947](https://github.com/planck-repl/planck/issues/947))
 
 ### Changed
 - `io/reader` on http follow redirects by default ([#937](https://github.com/planck-repl/planck/issues/937))
 - Update dependencies ([#940](https://github.com/planck-repl/planck/issues/940), [#941](https://github.com/planck-repl/planck/issues/941), [#942](https://github.com/planck-repl/planck/issues/942))
+- Use `paredit.js` for indentation ([#973](https://github.com/planck-repl/planck/issues/973))
+- Optimize `planck.io/copy` for input stream to writer case  
 - Promote `planck.socket.alpha` to `planck.socket` ([#956](https://github.com/planck-repl/planck/issues/956))
 
 ### Deprecated